Ajustanto retorno da rota /GET de transactions para retornar o balanço corretamente. Bug ligado a tipo incorreto na migration

+++ src/database/migrations/1592222390347-CreateTransaction.ts
@@ -21,7 +21,7 @@ export default class CreateTransaction1592222390347
           },
           {
             name: 'value',
-            type: 'decimal',
+            type: 'float',
             isNullable: false,
           },
           {
